requirements.txt
anthropic
tqdm
python-dotenv